Add snapcraft packaging
Use base: bare, with build-base: core24 for the smallest dependency set. At
runtime plz-run reaches out to systemd running on the host (boot base or
classic) so it does not require any other elements to be in place. The program
is built without cgo support, to avoid pulling in libc.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>